Revert "Refactoring spliting code"
