--- lib/hierarchical_db/base.rb
@@ -1,7 +1,0 @@
-require 'hierarchical_db/core'
-
-module HierarchicalDb
-  class Base
-    include Core
-  end
-end
